requirements.txt
fastapi
python-dotenv
livekit-agents[deepgram,groq,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
uvicorn